New C++11 header-only watershed implementation
This implementation follows the exact same logic found in the water.cpp
file, but implemented only using standard C++11 constructs and headers.
The implementation is offered as a header-only library to allow easy
reuse, and uses templates to let users provide their own interruption
mechanism for long-running problems.

The implementation requires the user to provide a pre-allocated
segmentation image that should be of the same size as the image to be
watershed. This is important for large images, in which case making
unnecessary copies of the data can exhaust the available resources and
lead to errors.

The code is not only more easily reusable, but also performs better.
This is due to: a) using std::vector instead of Rcpp's Vector classes,
since the former grows more nicely than the latter, b) using
pre-allocated vectors where previously new instances of a Vector would
be re-allocated each time, c) changing the order in which a few
conditions were checked for to allow short-circuiting them, and d)
changing the iteration order for neighboring pixels when watershedding a
particular "central" pixel.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>