Exposing new pure C++ watershed through Rcpp
We are calling this implementation "water_cpp_new", and exposing it
through the watershed='ProFound-new' option. The implementation is
simple: it creates the segment image, an Rcpp-based interrupt checker,
and passes everything down to our new C++11 watershed implementation.

Once the watershed algorithm is finished we need to add 1 to all
elements, since the C++11 code uses -1 to signal a no-segment cell,
while actual segments are numbered starting from 0. This is different
from the previous implementations, which use 0 to signal a no-segment
cell, and segments were numbered starting from 1.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>