New watershed is default, previous default is old
Since we now that our new C++11 implementation is value-equivalent to
the old implementation, and that it has better performance, we are
renaming things: 'water_cpp' now refers to the new implementation, while
'water_cpp_old' refers to the old. Likewise watershed='ProFound-new'
has been replaced with 'ProFound-old', which refers to 'water_cpp_old'.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>