Moving water.hpp -> water.h for CRAN compatibility
The "Writing R extensions" manual say they "We recommend using .h for
headers, also for C++", with a footnote sayind that "Using .hpp is not
guaranteed to be portable". This is a bit odd, as header files don't
really get compiled (only included), so I don't know where this
non-portability actually comes from.

Anyhow, this removes a warning out from the package checks.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>