Greatly simplifying akimaspline class
Firstly, the Initialise function is not needed if we simply define it as
the constructor of the class. Secondly, the isvalid member is not
needed. Thirdly, let's use member default initialisation. Fourthly,
there's no need for using double_t, let's go directly for double. And
finally, there's no need to keep declaration and definition separately,
this class is used only in this compilation unit.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>