Removing duplicated code
The different conditions to let values into the histogram were the only
small difference between how histograms were constructed, while the rest
was common code, but duplicated. This commit removes code duplication,
allowing the condition to be passed as an argument in order to implement
the three different logics.

On top of this I've simplified the code a bit: there was no need to have
a duplicate of the input "x" vector, I switched to using range-for
loops, and the code was in need of better spacing.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>