Removing duplicated code around sky RMS maths
Most of the calculations performed in the different cases supported by
the code (HI, LO, RHI, RLO, etc) was actually the same, but was a bit
difficult to distinguish when reading the code. This new version is more
readable (or so I think!), and thus should be easier to maintain.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>